+++ electron/database/database.ts
@@ -1,20 +1,20 @@
 import * as sqlite3 from 'sqlite3';
 import { ipcMain } from 'electron';
-import type { Project } from '../src/types/project';
-import type { Session } from '../src/types/session';
-import type { TagDatabase } from '../src/types/tag';
+import type { Project } from '../../src/types/project';
+import type { Session } from '../../src/types/session';
+import type { TagDatabase } from '../../src/types/tag';
 import type {
   UpdateResponse,
   DeleteResponse,
   DatabaseResponse,
-} from '../src/types/database-response';
+} from '../../src/types/database-response';
 import { getDatabaseConfig } from './database-config';
 import {
   getRecordAfterInsert,
   getRecordAfterWrite,
   getRecordBeforeDelete,
   setDatabaseInstance,
-} from './helpers';
+} from '../helpers';
 import { runMigrations } from './db-migrate';
 
 let db: sqlite3.Database;
