+++ src/main.tsx
@@ -2,7 +2,7 @@ import React, { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from '@/App.tsx';
-import { AppProvider } from '@/state/context/AppProvider';
+import { AppProvider } from '@/contexts/AppProvider';
 import { ThemeProvider } from 'styled-components';
 import { theme } from '@/styles/theme';
 import { DatabaseProvider } from '@/contexts/DatabaseContext';
